Restructure of footer almost done
