+++ package.json
@@ -7,6 +7,11 @@
     "@fluentui/react": "8.30.1",
     "@fluentui/react-hooks": "^8.2.4",
     "@fluentui/react-theme-provider": "^0.18.4",
+    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/free-brands-svg-icons": "^5.15.4",
+    "@fortawesome/free-regular-svg-icons": "^5.15.4",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@fortawesome/react-fontawesome": "^0.1.15",
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.11.2",
     "@testing-library/jest-dom": "5.11.9",
