+++ src/components/Footer.tsx
@@ -1,51 +1,110 @@
-import { Link, Icon } from 'office-ui-fabric-react';
+import { Link } from 'office-ui-fabric-react';
 import { Text } from 'office-ui-fabric-react/lib/Text';
 import { Container } from 'react-bootstrap';
 import { semibold } from '../services/fonts';
 import { useTheme } from '@fluentui/react-theme-provider';
-import { withCookies } from 'react-cookie';
+import { useCookies, withCookies } from 'react-cookie';
+import { IDropdownOption, Dropdown, IIconProps, PrimaryButton, Toggle, TooltipHost, IconButton, SwatchColorPicker, ITooltipHostStyles } from '@fluentui/react';
+import { palettes } from '../services/palettes';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { addDays } from '../services/Utils';
+import { library } from '@fortawesome/fontawesome-svg-core'
+import { fab } from '@fortawesome/free-brands-svg-icons'
+import { faCommentDots } from '@fortawesome/free-solid-svg-icons'
 import Col from 'react-bootstrap/Col';
 import Row from 'react-bootstrap/Row';
 import LocalizationService from "../services/LocalizationService";
-import JsxParser from 'react-jsx-parser';
 
+library.add(fab, faCommentDots);
 const listElement = { marginBottom: '.2rem' };
 
-const Footer = () => {
+const footerIcons: any = [
+    { name: { it: 'Canale Telegram', en: 'Telegram Channel' }, link: 'https://t.me/studenti_unimi', iconName: 'telegram-plane', type: 'brand' },
+    { name: { it: 'Gruppo Principale', en: 'Main Group' }, link: 'https://t.me/unimichat', iconName: 'comment-dots', type: 'normal' },
+    { name: { it: 'Canale Discord', en: 'Discord Channel' }, link: 'https://discord.gg/SwPzAkv4A4', iconName: 'discord', type: 'brand' },
+    { name: { it: 'Organizzazione GitHub', en: 'GitHub Organization' }, link: 'https://github.com/StudentiUnimi', iconName: 'github', type: 'brand' },
+    { name: { it: 'Pagina Facebook', en: 'Facebook Page' }, link: 'https://www.facebook.com/networkstudentiunimi', iconName: 'facebook', type: 'brand' },
+    { name: { it: 'Pagina Instagram', en: 'Instagram Page' }, link: 'https://www.instagram.com/studentiunimi.it/', iconName: 'instagram', type: 'brand' },
+];
+
+interface Props { changeTheme: () => void, changePalette: (id: string) => void };
+
+const Footer = (props: Props) => {
     var theme = useTheme();
+    const [cookies, setCookie] = useCookies();
+    const date: Date = addDays(new Date(), 90);
+
+    const themeToggled = () => {
+        if (cookies["theme"] === "dark") setCookie("theme", "light", { path: "/", expires: date });
+        else { setCookie("theme", "dark", { path: "/", expires: date }); }
+        props.changeTheme();
+    };
+
+    if (cookies["theme"] === undefined) {
+        /*
+        // This is theme initialization based on browser; this isn't working, so I must fix this.
+        setCookie("theme", "light", { path: "/", expires: date }); 
+        const darkThemeMq = window.matchMedia("(prefers-color-scheme: dark)");
+        if (darkThemeMq.matches) themeToggled();
+        */
+        setCookie("theme", "light", { path: "/", expires: date }); 
+    }
+    
+    if (cookies['language'] === undefined) { 
+        const isNavLanguageITA = isNavigatorLanguageItalian();
+        setCookie("language", (isNavLanguageITA ? 'it' : 'en'), { path: "/", expires: date }); 
+    }
+
+    LocalizationService.localize(cookies['language']);
     const locale = LocalizationService.strings();
-    const iconStyle = { fontSize: 18 };
+    var language: string = LocalizationService.getLanguage();
+
+    const changeLanguage = (key: string) => {
+        LocalizationService.localize(key);
+    };
+
+    const languageOptions: IDropdownOption[] = [
+        { key: 'it', text: locale.settingsPanel.italian },
+        { key: 'en', text: locale.settingsPanel.english }
+    ];
+
+    if (cookies["palette"] === undefined) 
+    { 
+        setCookie("palette", "a", { path: "/", expires: date });
+    }
+    
+    const wrapIconStyle = { backgroundColor: theme.palette.themeSecondary, borderRadius: '50%', minWidth: 35, minHeight: 35, display: 'inline-block', textAlign: 'center', justifyContent: 'center', verticalAlign: 'middle' } as React.CSSProperties;
+    const iconStyle = { color: theme.palette.white, fontSize: '22px', marginTop: 6 };
+
+    const buttonStyle = { maxWidth: '250px', boxShadow: theme.effects.elevation8 };
+    const buttonIconProps: IIconProps = { iconName: 'ChevronRightSmall', styles: { root: { fontSize: 12 } } };
+
+    /* Theme palette code */
+    const colorCells: any[] = palettes.map(x => ({ id: x.id, label: x.label, color: x.palette?.themePrimary }));
+    const resetColorIcon: IIconProps = { iconName: 'SyncOccurence' };
+    const calloutPropsResetColor = { gapSpace: 10 };
+    const hostStylesResetColor: Partial<ITooltipHostStyles> = { root: { display: 'inline-block' } };
+
     return (
         <footer style={{ backgroundColor: theme.palette.neutralQuaternaryAlt, borderTop: '1px solid', borderColor: theme.palette.neutralLight }} className="pt-4 pb-4">
             <Container style={{ width: '100%', color: theme.palette.neutralSecondary }}>
 
-                <Row>
-                    <Col xl={4} lg={4} md={4} sm={12} xs={12} className="mb-4 mb-md-0">
+                <Row className="mb-4">
+                    <Col xl={4} lg={4} md={6} sm={12} xs={12} className="mb-4 mb-md-0">
                         <div className="mb-2">
-                            <Text styles={semibold} variant="medium">
-                                <Link href="https://github.com/StudentiUnimi" target="_blank">&copy; Network StudentiUniMi</Link>
+                            <Text styles={semibold} style={{ color: theme.palette.themePrimary }} variant="large">
+                                &copy; Network StudentiUniMi
                             </Text>
                         </div>
 
                         <div className="mb-2 text">
-                            <Text variant="medium">
-                                {locale.footer[0].text}
-                            </Text>
-                        </div>
-
-                        <div className="mb-1">
-                            <Link href="https://t.me/studenti_unimi" target="_blank" className="text-decoration-none mr-1"><Icon iconName="Send" style={iconStyle}/></Link>
-                            <Link href="https://t.me/unimichat" target="_blank" className="text-decoration-none mr-1"><Icon iconName="Group" style={iconStyle} /></Link>
-                            <Link href="https://discord.gg/SwPzAkv4A4" target="_blank" className="text-decoration-none mr-1"><Icon iconName="Game" style={iconStyle}/></Link>
-                            <Link href="https://github.com/StudentiUnimi/website" target="_blank" className="text-decoration-none mr-1"><Icon iconName="Website" style={iconStyle}/></Link>
-                            <Link href="https://github.com/StudentiUniMi/backend" target="_blank" className="text-decoration-none mr-1"><Icon iconName="ServerEnviroment" style={iconStyle} /></Link>
-                            <Link href="https://www.facebook.com/networkstudentiunimi" target="_blank" className="text-decoration-none"><Icon iconName="PageLink" style={iconStyle}/></Link>
+                            <PrimaryButton text="Entra nel nostro gruppo" iconProps={buttonIconProps} href="https://t.me/unimichat" target="_blank" className="text-decoration-none" allowDisabledFocus style={buttonStyle} />
                         </div>
                     </Col>
                     
 
-                    <Col xl={4} lg={4} md={4} sm={12} xs={12} className="mb-4 mb-md-0">
-                        <div className="mid-column">
+                    <Col xl={2} lg={2} md={3} sm={12} xs={12} className="mb-2 mb-md-0">
+                        <div /*className="mid-column"*/>
                             <div className="mb-2">
                                 <Text styles={semibold} variant="medium">{locale.footer[1].header}</Text>
                             </div>
@@ -54,17 +113,13 @@ const Footer = () => {
                                 <Text variant="medium">
                                     <ul className="list-unstyled mb-3">
                                         <li style={listElement}>
-                                            <Link href="http://www.quickunimi.it/" target="_blank"><Icon iconName="Globe" /> QuickUnimi</Link>
+                                            <Link href="http://www.quickunimi.it/" target="_blank">QuickUnimi</Link>
                                         </li>
                                         <li style={listElement}>
-                                            <Link href="https://orientamento.di.unimi.it/index.php/studia-con-noi/tutor-di-processo" target="_blank"><Icon iconName="SunQuestionMark" /> FAQ matricole</Link>
+                                            <Link href="https://quanto-manca.it/" target="_blank">Quanto-manca.it</Link>
                                         </li>
                                         <li style={listElement}>
-                                            <Link href="https://quanto-manca.it/" target="_blank"><Icon iconName="DateTime" /> Quanto-manca.it</Link>
-
-                                        </li>
-                                        <li style={listElement}>
-                                            <Link href="https://codeshare.tech" target="_blank"><Icon iconName="FileCode" /> Codeshare.tech</Link>
+                                            <Link href="https://codeshare.tech" target="_blank">Codeshare.tech</Link>
                                         </li>
                                     </ul>
                                 </Text>
@@ -72,18 +127,78 @@ const Footer = () => {
                         </div>
                     </Col>
 
-                    <Col xl={4} lg={4} md={4} sm={12} xs={12} className="mb-4 mb-md-0 contacts">
-                        <div className="mb-2">
-                            <Text styles={semibold} variant="medium">{locale.footer[2].header}</Text>
+                    <Col xl={2} lg={2} md={3} sm={12} xs={12}>
+                        <div>
+                            <Toggle
+                                label={locale.settingsPanel.changeTheme}
+                                onText={locale.settingsPanel.darkTheme}
+                                offText={locale.settingsPanel.lightTheme}
+                                checked={cookies["theme"] === "dark"}
+                                onChange={themeToggled}
+                                theme={theme}
+                            />
                         </div>
+                        
+                        <div>
+                            <Dropdown
+                                label={locale.settingsPanel.selectLanguage}
+                                options={languageOptions}
+                                selectedKey={cookies["language"]}
+                                onChange={(_, option) => { changeLanguage(option!.key as string); setCookie("language", option!.key as string, { path: "/", expires: date }) }}
+                                theme={theme}
+                                style={{ maxWidth: 200 }}
+                            />
+                        </div>
+                    </Col>
 
-                        <div className="mb-1 text">
-                            <Text variant="medium">
-                                <JsxParser bindings={{ theme: theme }} components={{ Text, Link }} jsx={locale.footer[2].text} />
-                            </Text>
+                    <Col xl={4} lg={4} md={6} sm={12} xs={12}>
+                        <div>
+                            <Text variant="medium" styles={semibold}>{locale.settingsPanel.selectColor}  </Text>
+                            <TooltipHost
+                                content="Reset color"
+                                calloutProps={calloutPropsResetColor}
+                                styles={hostStylesResetColor}
+                            >
+                                <IconButton iconProps={resetColorIcon} onClick={() => { setCookie("palette", 'a', { path: "/", expires: date }); props.changePalette('a'); }} />
+                            </TooltipHost>
+                            <SwatchColorPicker selectedId={cookies["palette"]} columnCount={10} cellShape={'square'} colorCells={colorCells} onColorChanged={(id) => { setCookie("palette", id, { path: "/", expires: date }); props.changePalette(id!); }} />
                         </div>
                     </Col>
                 
+                </Row>
+
+
+                <Row>
+                    <Col lg={6} sm={12} style={{ display: 'table' }} className="center-mobile mb-2">
+                        <Text variant="medium"  style={{  display: 'table-cell', verticalAlign: 'middle' }}>
+                            {locale.footer[0].text}
+                        </Text>
+                    </Col>
+
+                    <Col lg={6} sm={12}>
+                        <div className="mb-1 text-right center-mobile">
+                            {footerIcons.map( (x: any, i: number) => { 
+                                return (
+                                    <TooltipHost
+                                        content={x.name[language]}
+                                        calloutProps={calloutPropsResetColor}
+                                        styles={hostStylesResetColor}
+                                        key={i}
+                                    >
+                                        <Link href={x.link} target="_blank">
+                                            <span style={wrapIconStyle} className="text-decoration mr-1">
+                                                { x.type === 'brand' ?
+                                                <FontAwesomeIcon icon={['fab', x.iconName]} style={iconStyle} />
+                                                :
+                                                <FontAwesomeIcon icon={x.iconName} style={iconStyle} /> }
+                                            </span>
+                                        </Link>
+                                    </TooltipHost>
+                                )}
+                            )}
+                        </div>
+                    </Col>
+
                 </Row>
 
             </Container>
@@ -91,4 +206,16 @@ const Footer = () => {
     )
 };
 
-export default withCookies(Footer);
+export default Footer;
+
+/**
+ * This function returns true if the navigator language is italian.
+ */
+ const isNavigatorLanguageItalian = () => {
+    const navLanguage = navigator.language;
+    if (navLanguage === 'it') return true;
+
+    const s: string[] = navLanguage.split("-",2);
+    if (s.length >= 2 && s[0] === 'it') return true;
+    return false;
+}
